requirements.txt
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
idna==2.9
python-dateutil==2.8.1
regex==2020.2.20
requests==2.23.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
